chore: cache terraform providers for workspaces terraform tests
